Add picture upload by user
Users can now upload an image when creating a new list.
If a picture is joined, it is displayed as the list background on the home page.
It is also displayed as the main background picture when looking at the details of a list.

The pictures are centered on faces when there are any